+++ fasp/search/__init__.py
@@ -1,3 +1,3 @@
-from fasp.search.discovery_search_client import DataConnectClient
+from fasp.search.data_connect_client import DataConnectClient
 from fasp.search.BigQuerySearchClient import BigQuerySearchClient
 from fasp.search.local_search import Gen3ManifestClient
